requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv==1.0.0
aiohttp>=3.8.0
backoff==2.2.1
asyncio>=3.4.3
fal-client==0.7.0 